strawberry-django-demo/pyproject.toml
[tool.poetry]
name = "strawberry-django-demo"
version = "0.1.0"
description = ""
authors = ["kuroyasu <xjr1300.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql-django = "^0.47.1"
django-choices-field = "^2.3.0"


[tool.poetry.group.dev.dependencies]
django-stubs = { extras = ["compatible-mypy"], version = "^5.0.4" }
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
